finished part4

+++ part4/blog_list/controllers/blogs.js
@@ -1,10 +1,13 @@
 const blogRouter = require("express").Router();
 const res = require("express/lib/response");
 const Blog = require("../models/blog");
+const User = require("../models/user");
 const { info } = require("../utils/logger");
+const jwt = require("jsonwebtoken");
+const { userExtractor } = require("../utils/middleware");
 
 blogRouter.get("/", async (request, response) => {
-  const blogs = await Blog.find({});
+  const blogs = await Blog.find({}).populate("user", { username: 1, name: 1 });
   response.json(blogs);
 });
 
@@ -17,19 +20,27 @@ blogRouter.get("/:id", async (request, response, next) => {
   }
 });
 
-blogRouter.post("/", async (request, response, next) => {
+blogRouter.post("/", userExtractor, async (request, response, next) => {
   const body = request.body;
+  const user = request.user;
+  info("the controller is getting the user:", user);
+
   if (!body.title || !body.url) {
     response.status(400).send("Bad request");
+  } else if (!user) {
+    response.status(401).send("invalid token");
   } else {
     const blog = new Blog({
       title: body.title,
-      author: body.author,
+      author: user.name,
       likes: body.likes || 0,
       url: body.url,
+      user: user._id,
     });
     try {
       const savedBlog = await blog.save();
+      user.blogs = user.blogs.concat(savedBlog._id);
+      await user.save();
       response.status(201).json(savedBlog);
     } catch (exception) {
       next(exception);
@@ -37,10 +48,14 @@ blogRouter.post("/", async (request, response, next) => {
   }
 });
 
-blogRouter.delete("/:id", async (request, response, next) => {
+blogRouter.delete("/:id", userExtractor, async (request, response, next) => {
   try {
-    await Blog.findByIdAndDelete(request.params.id);
-    response.status(204).end();
+    const blog = await Blog.findById(request.params.id);
+    const user = request.user;
+    if (blog.user.toString() === user.id.toString()) {
+      await Blog.findByIdAndDelete(request.params.id);
+      response.status(204).end();
+    }
   } catch (exception) {
     next(exception);
   }
